+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: pscott <pscott@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 12:39:50 by pscott            #+#    #+#             */
-/*   Updated: 2018/11/07 12:42:17 by pscott           ###   ########.fr       */
+/*   Updated: 2018/11/07 16:10:42 by pscott           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int i;
+	size_t i;
 
 	i = ft_strlen(s) - 1;
 	while (i > -1)
